+++ idearly/src/pages/LoginPage/components/LoginVaildationInput/LoginVaildationInput.types.ts
@@ -1,0 +1,13 @@
+import { UseFormRegisterReturn } from "react-hook-form";
+
+export interface IValidationInput {
+  label: string;
+  type: string;
+  placeholder: string;
+  errors:
+    | {
+        message?: string;
+      }
+    | undefined;
+  register: UseFormRegisterReturn;
+}
